requirements.txt
ipykernel
pandas==2.2.3
numpy==2.2.6
ensure==1.0.4
python-box==7.3.2
PyYAML==6.0.2
python-dotenv==1.1.0
psycopg2==2.9.10
scipy==1.15.3
scikit-learn==1.7.0